requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-jsonfield==1.4.1
django-log-request-id==2.0.0
django-redis==5.2.0
django-requestlogs==0.7.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
JSON-log-formatter==0.5.1
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2022.2.1
redis==3.5.3
redis-py-cluster==2.1.3
six==1.16.0
sqlparse==0.4.2
wrapt==1.14.1
